Added more questions to questions.json + other things.
I added more questions to questions.json and added a difficulty property
to each question in preparation for the algorithms implementation. (#73)

I also added an applause sound which plays when a user completes a quiz
or test. However, It's a little weird to hear people clapping after you
get 0% but for right now it doesn't matter that much.

This commit supports the resolution to the problem described in Support Story #72.