Implemented a question recommendation algorithm (for lesson quizzes).
We recognized the importance of reducing difficulty if a user started to struggle on their quiz.
We also recognized the importance of challenging the more knowledgeable users by increasing their question difficulty.

This recommendation algorithm is designed to serve easier or harder questions depending on how the user is doing on the quiz at a given point.

pros:
    - questions are not completely randomly served.
    - questions can get easier or harder depending on a users progress.
    - opens the door to understanding the users relative comprehension of a concept in the future.

cons:
    - requires a lot of questions. a lot.
    - implementation of the algorithm is very weird.
    - made the backend an even bigger mess.

However, I'm happy with it.

Now all there is to do is to actually implement the part where it looks at your current progress because at this current momment it is simulated by a random boolean.

This commit supports the resolution to the problem described in Support Story #72 and Task #73.