Fixed an annoying bug with the algorithms implementation.
There was a weird bug with the quiz system that repeated the first
question in the previous commit.

I am 90% certain I fixed it but I don't know for sure if the fix
actually fixed it or even if it created another problem I am unaware of.

Until I am certain, I will not merge this branch (algorithm) into dev.