Added a "try except" block to handle a potential error with quizzes.
So, A few commits back I said I fixed an issue with the algorithm.
Well turns out, I kind of didn't.

The issue was with the amount of questions in questions.json having less
questions than expected. By adding a "try catch" block to the code that
could cause this issue, We can better understand the reason why the code
isn't actually working.

This resolves the issue described in Support Story #72 and Task #73.