Removed line that created an issue with accounts.
There was an issue with the project getting mad if you didn't include a last name.

Although I was unable to replicate the issue described in task #76, I removed the line that caused that such Error (it isn't used anymore any anyway).

I will be marking this task as Complete but keeping te Issue open in case I didn't actually fix it.