Created a proper "Congratulations" page.
Prior this commit when you finished the entire course you were simply given this message:

> Congratulations! You have completed all the lessons.

Which is... uneventful and quite boring. Which is why I made a custom webpage for it.

It will be improved after MVP but for now, it's good enough.

This fixes the issue described in Support Story #86 and Task #87.