Commited Inital Prototype for settings page.
To prepare for MVP, We need to have a nice settings page so the user can change the
associated account email and password as well as selecting a primary language they
will use.

This is the first commit in a seris of commits for a settings page.

In this commit I added:

1) A Responisve Navigation Menu
2) A Settings Customization View (I don't know what else to call it)